+++ Examples/EndToEndExample/10-OngoingConfiguration.ps1
@@ -277,15 +277,15 @@ Configuration Example
         }
 
         ###Mailbox Server settings###
-        $dbMap          = DBMapFromServersCsv -ServersCsvPath $Node.ServersCsvPath `
+        $dbMap          = Get-DBMapFromServersCsv -ServersCsvPath $Node.ServersCsvPath `
                                               -ServerNameInCsv $Node.ServerNameInCsv `
                                               -DbNameReplacements $dagSettings.DbNameReplacements
 
-        $primaryDbList  = DBListFromMailboxDatabasesCsv -MailboxDatabasesCsvPath $Node.MailboxDatabasesCsvPath `
+        $primaryDbList  = Get-DBListFromMailboxDatabasesCsv -MailboxDatabasesCsvPath $Node.MailboxDatabasesCsvPath `
                                                         -ServerNameInCsv $Node.ServerNameInCsv `
                                                         -DbNameReplacements $dagSettings.DbNameReplacements
 
-        $copyDbList     = DBListFromMailboxDatabaseCopiesCsv -MailboxDatabaseCopiesCsvPath $Node.MailboxDatabaseCopiesCsvPath `
+        $copyDbList     = Get-DBListFromMailboxDatabaseCopiesCsv -MailboxDatabaseCopiesCsvPath $Node.MailboxDatabaseCopiesCsvPath `
                                                              -ServerNameInCsv $Node.ServerNameInCsv `
                                                              -DbNameReplacements $dagSettings.DbNameReplacements